+++ backend/src/ee/services/audit-log/audit-log-types.ts
@@ -6736,6 +6736,8 @@ interface CreateOauthClientEvent {
     clientId: string;
     name: string;
     grantTypes: string[];
+    tokenExchangeAudience?: string | null;
+    tokenExchangeIdpSatisfiesMfa: boolean;
   };
 }
