+++ frontend/src/hooks/api/oauthClients/index.tsx
@@ -7,3 +7,4 @@ export {
 } from "./mutations";
 export { useGetOauthAuthorizeInfo, useGetOauthClients } from "./queries";
 export type { TOauthAuthorizeInfo, TOauthClient } from "./types";
+export { OauthGrantType } from "./types";
